saving work

+++ testsuites/CBLTester/conftest.py
@@ -2,6 +2,7 @@ import time
 import pytest
 
 from utilities.cluster_config_utils import persist_cluster_config_environment_prop
+from keywords.constants import SDK_TIMEOUT
 from keywords.utils import log_info
 from keywords.utils import host_for_url
 from keywords.ClusterKeywords import ClusterKeywords
@@ -9,6 +10,7 @@ from keywords.couchbaseserver import CouchbaseServer
 from keywords.constants import CLUSTER_CONFIGS_DIR
 from keywords.MobileRestClient import MobileRestClient
 from keywords.SyncGateway import sync_gateway_config_path_for_mode
+from keywords.SyncGateway import SyncGateway
 from keywords.exceptions import ProvisioningError
 from keywords.tklogging import Logging
 from CBLClient.Replicator_new import Replicator
@@ -65,6 +67,11 @@ def pytest_addoption(parser):
                      action="store_true",
                      help="xattrs: Enable xattrs for sync gateway")
 
+    parser.addoption("--create-db-per-test",
+                     action="store_true",
+                     help="create-db-per-test: Creates/deletes client DB for every test. \
+                     Default is to create/delete client DB per suite")
+
     parser.addoption("--no-conflicts",
                      action="store_true",
                      help="If set, allow_conflicts is set to false in sync-gateway config")
@@ -88,8 +95,11 @@ def params_from_base_suite_setup(request):
     server_version = request.config.getoption("--server-version")
     enable_sample_bucket = request.config.getoption("--enable-sample-bucket")
     xattrs_enabled = request.config.getoption("--xattrs")
+    create_db_per_test = request.config.getoption("--create-db-per-test")
+
     base_url = "http://{}:{}".format(liteserv_host, liteserv_port)
     cluster_config = "{}/base_{}".format(CLUSTER_CONFIGS_DIR, mode)
+    sg_config = sync_gateway_config_path_for_mode("sync_gateway_travel_sample", mode)
     no_conflicts_enabled = request.config.getoption("--no-conflicts")
     cluster_utils = ClusterKeywords()
     cluster_topology = cluster_utils.get_cluster_topology(cluster_config)
@@ -132,7 +142,6 @@ def params_from_base_suite_setup(request):
         log_info("Running with allow conflicts")
         persist_cluster_config_environment_prop(cluster_config, 'no_conflicts_enabled', False)
 
-    sg_config = sync_gateway_config_path_for_mode("sync_gateway_travel_sample", mode)
     cluster_utils = ClusterKeywords()
     cluster_topology = cluster_utils.get_cluster_topology(cluster_config)
     cbs_url = cluster_topology['couchbase_servers'][0]
@@ -153,34 +162,64 @@ def params_from_base_suite_setup(request):
             logging_helper.fetch_and_analyze_logs(cluster_config=cluster_config, test_name=request.node.name)
             raise
 
+    if enable_sample_bucket and create_db_per_test:
+        raise Exception("Invalid options combination")
+
+    source_db = None
+    if not create_db_per_test:
+        # Create CBL database
+        cbl_db = "test_db"
+        db = Database(base_url)
+
+        log_info("Creating a Database {} at the suite setup".format(cbl_db))
+        source_db = db.create(cbl_db)
+        log_info("Getting the database name")
+        db_name = db.getName(source_db)
+        assert db_name == "test_db"
+
     if enable_sample_bucket:
         server_url = cluster_topology["couchbase_servers"][0]
         server = CouchbaseServer(server_url)
 
-        server.delete_buckets()
-        time.sleep(5)
+        buckets = server.get_bucket_names()
+        if enable_sample_bucket in buckets:
+            log_info("Deleting existing {} bucket".format(enable_sample_bucket))
+            server.delete_bucket(enable_sample_bucket)
+            time.sleep(5)
+
+        log_info("Loading sample bucket {}".format(enable_sample_bucket))
         server.load_sample_bucket(enable_sample_bucket)
         server._create_internal_rbac_bucket_user(enable_sample_bucket)
-        time.sleep(180)
+
+        # Restart SG after the bucket deletion
+        sync_gateways = cluster_topology["sync_gateways"]
+        sg_obj = SyncGateway()
+
+        for sg in sync_gateways:
+            sg_ip = host_for_url(sg["admin"])
+            log_info("Restarting sync gateway {}".format(sg_ip))
+            sg_obj.restart_sync_gateways(cluster_config=cluster_config, url=sg_ip)
+            time.sleep(5)
+
+        if mode == "di":
+            ac_obj = SyncGateway()
+            sg_accels = cluster_topology["sg_accels"]
+            for ac in sg_accels:
+                ac_ip = host_for_url(ac)
+                log_info("Restarting sg accel {}".format(ac_ip))
+                ac_obj.restart_sync_gateways(cluster_config=cluster_config, url=ac_ip)
+                time.sleep(5)
 
         # Create primary index
+        password = "password"
+        log_info("Connecting to {}/{} with password {}".format(cbs_ip, enable_sample_bucket, password))
+        sdk_client = Bucket('couchbase://{}/{}'.format(cbs_ip, enable_sample_bucket), password=password, timeout=SDK_TIMEOUT)
         log_info("Creating primary index for {}".format(enable_sample_bucket))
         sdk_client = Bucket('couchbase://{}/{}'.format(cbs_ip, enable_sample_bucket), password='password')
         n1ql_query = 'create primary index on {}'.format(enable_sample_bucket)
         query = N1QLQuery(n1ql_query)
         sdk_client.n1ql_query(query)
 
-    # Create CBL database
-    cbl_db = "test_db"
-    db = Database(base_url)
-
-    log_info("Creating a Database {}".format(cbl_db))
-    source_db = db.create(cbl_db)
-    log_info("Getting the database name")
-    db_name = db.getName(source_db)
-    assert db_name == "test_db"
-
-    if enable_sample_bucket:
         # Start continuous replication
         repl_obj = Replicator(base_url)
         auth_obj = BasicAuthenticator(base_url)
@@ -190,7 +229,9 @@ def params_from_base_suite_setup(request):
                                         replication_type="PUSH_AND_PULL",
                                         replicator_authenticator=authenticator)
         repl_obj.start(replicator)
-        time.sleep(60)
+        # Wait for replication to complete
+        # TODO Wait for replication state idle
+        time.sleep(300)
 
     yield {
         "cluster_config": cluster_config,
@@ -209,6 +250,7 @@ def params_from_base_suite_setup(request):
         "target_admin_url": target_admin_url,
         "base_url": base_url,
         "enable_sample_bucket": enable_sample_bucket,
+        "create_db_per_test": create_db_per_test,
         "source_db": source_db,
         "cbl_db": cbl_db
     }
@@ -217,17 +259,20 @@ def params_from_base_suite_setup(request):
         log_info("Stopping replication")
         repl_obj.stop(replicator)
 
-    # Delete CBL database
-    log_info("Deleting the database {}".format(cbl_db))
-    db.deleteDB(source_db)
+    if not create_db_per_test:
+        # Delete CBL database
+        log_info("Deleting the database {} at the suite teardown".format(cbl_db))
+        db.deleteDB(source_db)
 
 
 @pytest.fixture(scope="function")
 def params_from_base_test_setup(request, params_from_base_suite_setup):
+    base_url = params_from_base_suite_setup["base_url"]
     cluster_config = params_from_base_suite_setup["cluster_config"]
     xattrs_enabled = params_from_base_suite_setup["xattrs_enabled"]
     liteserv_host = params_from_base_suite_setup["liteserv_host"]
     liteserv_port = params_from_base_suite_setup["liteserv_port"]
+    create_db_per_test = params_from_base_suite_setup["create_db_per_test"]
     no_conflicts_enabled = params_from_base_suite_setup["no_conflicts_enabled"]
     target_admin_url = params_from_base_suite_setup["target_admin_url"]
     source_db = params_from_base_suite_setup["source_db"]
@@ -251,6 +296,16 @@ def params_from_base_test_setup(request, params_from_base_suite_setup):
     log_info("mode: {}".format(mode))
     log_info("xattrs_enabled: {}".format(xattrs_enabled))
 
+    if create_db_per_test:
+        # Create CBL database
+        db = Database(base_url)
+
+        log_info("Creating a Database {} at test setup".format(cbl_db))
+        source_db = db.create(cbl_db)
+        log_info("Getting the database name")
+        db_name = db.getName(source_db)
+        assert db_name == "test_db"
+
     # This dictionary is passed to each test
     yield {
         "cluster_config": cluster_config,
@@ -271,9 +326,17 @@ def params_from_base_test_setup(request, params_from_base_suite_setup):
         "cbl_db": cbl_db,
     }
 
+    if create_db_per_test:
+        # Delete CBL database
+        log_info("Deleting the database {} at test teardown".format(cbl_db))
+        db.deleteDB(source_db)
+
+
 @pytest.fixture(scope="class")
 def class_init(request, params_from_base_suite_setup):
     base_url = params_from_base_suite_setup["base_url"]
+    liteserv_platform = params_from_base_suite_setup["liteserv_platform"]
+
     db_obj = Database(base_url)
     doc_obj = Document(base_url)
     dict_obj = Dictionary(base_url)
@@ -296,6 +359,8 @@ def class_init(request, params_from_base_suite_setup):
     request.cls.sg_client = sg_client
     request.cls.db_obj = db_obj
     request.cls.db = db
-    yield 
+    request.cls.liteserv_platform = liteserv_platform
+
+    yield
     db_obj.deleteDB(db)
